backend/requirements.txt
fastapi
uvicorn
pydantic
asyncpg
python-dotenv
email-validator
passlib
argon2-cffi
python-jose
python-multipart
python-dateutil
